refactor: share Song/Requester types and flatten IIFE in request.ts
Dedupe the Song/Track shape that was declared in four places into a new
shared/types.ts module, following the precedent of shared/formFields.ts.
Flatten the awkward IIFE wrapping deriveFormResponseConfig() in request.ts
into a plain try/catch at the handler's natural order.

Changes:
- Add shared/types.ts exporting Song and Requester
- src/hooks/useSongSearch.ts re-exports Song from shared
- src/lib/googleForm.ts imports Song/Requester from shared; drop the local
  SongRequestDetails alias
- netlify/functions/search.ts uses Song[] in SearchResponse
- netlify/functions/request.ts replaces the IIFE+Error check with a direct
  try/catch returning 500 on config errors

Type-only refactor — no behaviour change. All 54 unit tests pass, Playwright
smoke passes, lint clean, build succeeds.

Closes #33
