+++ docs/plan/issues/033_refactor_share_song_types_dedupe_derive_form_config.md
@@ -1,7 +1,7 @@
 # GitHub Issue #33: Refactor — share Song/Track types and dedupe deriveFormResponseConfig call
 
 **Issue:** [#33](https://github.com/denhamparry/djrequests/issues/33)
-**Status:** Reviewed (Approved)
+**Status:** Complete
 **Date:** 2026-04-16
 
 ## Problem Statement
